split analysis and plotting; analysis is done per transformation
